# public/swagger/spec-files/default.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: MyPet API
  description: >
    La API que expone el servidor MyPet
host: localhost:8080
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /user:
    get:
      description: Información de usuario
      parameters:
        - $ref: "#/parameters/Token"
      responses:
        200:
          description: Datos de usuario
          schema:
            type: object
            properties:
              name:
                type: string
    post:
      description: Crear un usuario
      parameters:
        - description: Datos de usuario
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UserData"
      responses:
        200:
          $ref: "#/responses/Success"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
    put:
      description: Modificar un usuario
      parameters:
        - description: Datos de usuario
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UserData"
        - $ref: "#/parameters/Token"
      responses:
        200:
          $ref: "#/responses/Success"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
...
definitions:
  UserData:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
        format: password
...
  Error:
    type: object
    properties:
      error:
        type: object
        properties:
          description:
            type: string
            description: Representación en texto del error
          code:
            type: integer
            description: Código de error
parameters:
  Token:
    description: Token de autenticación
    in: query
    name: token
    type: string
    required: true
responses:
  Success:
    description: Acción realizada correctamente
    schema:
      type: object
      properties:
        success:
          description: Mensaje
  Unauthorized:
    description: Error de autenticación
    schema:
      $ref: "#/definitions/Error"
  BadRequest:
    description: No se especificaron todos los parametros necesarios
    schema:
      $ref: "#/definitions/Error"